Moving constants to proper file
