+++ src/ColumnSort/index.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import userEvent from '@testing-library/user-event';
 import { within } from '@testing-library/dom';
 
-import { SORT_DIRECTION } from '../ducks/event-filter';
+import { SORT_DIRECTION } from '../constants';
 import { EVENT_SORT_ORDER_OPTIONS } from '../utils/event-filter';
 import ColumnSort from './';
 import { render, screen } from '../test-utils';
