+++ src/EventFilter/index.js
@@ -9,7 +9,8 @@ import debounce from 'lodash/debounce';
 import noop from 'lodash/noop';
 
 import { BREAKPOINTS } from '../constants';
-import { updateEventFilter, INITIAL_FILTER_STATE, DEFAULT_EVENT_SORT } from '../ducks/event-filter';
+import { updateEventFilter, INITIAL_FILTER_STATE } from '../ducks/event-filter';
+import { DEFAULT_EVENT_SORT } from '../constants';
 import { isFilterModified } from '../utils/event-filter';
 import { resetGlobalDateRange } from '../ducks/global-date-range';
 import { trackEventFactory, EVENT_FILTER_CATEGORY, REPORTS_CATEGORY } from '../utils/analytics';
